Fix a performance regression in retrieving parse data
R has a bug for packages where it does not attach parse data to anything
but the last file in the package. So to get the parse data for the rest
of the files you need to re-parse each file from the package database.

The previous code did this re-parsing every time the parse data was
missing, which was extremely slow if the package had a lot of code. We
now parse this once for the package and store the parsed results, fixing
the performance regression.

Fixes #274