Fix cases where the expression is on the actual rather than parsed line
I am not entirely sure when this happens, the srcref help says the only
time parsed line and actual line are different are when line directives
are used. R uses line directives in packages with more than one source
file, because it combines all source files into one file. So perhaps
this effects any source file in a package which is not the first?

Why we can't use the line starts and ends unconditionally I am not sure,
but it seems like for the first file we actually want the parsed
line...

The R parser / getParseData has no end of corner cases.

Fixes https://github.com/r-lib/covr/issues/317