Avoid cyclic import for httr
We changed to using an importFrom directive so we could mock these
functions, but doing that makes a cyclic dependency if you try to
calculate coverage on httr and its dependencies, because they are then
load time dependencies for covr. Here we define a simple pass through
function to make these functions runtime dependencies instead, which
avoids the issue as they will only be loaded after httr itself is
loaded.

Fixes #623